rename EditTeam.svelte to TeamEdit
