+++ html/src/routes.ts
@@ -8,7 +8,7 @@ import MyTeam from "./lib/pages/MyTeam.svelte";
 import Invite from "./lib/pages/Invite.svelte";
 import UserTeams from "./lib/pages/UserTeams.svelte";
 import ProfilePage from "./lib/pages/ProfilePage.svelte";
-import EditTeam from "./lib/pages/EditTeam.svelte";
+import TeamEdit from "./lib/pages/TeamEdit.svelte";
 
 export default {
     '/': HomePage,
@@ -22,5 +22,5 @@ export default {
     '/teams/browse': TeamsBrowse,
     '/teams/create': TeamsCreate,
     '/profile': ProfilePage,
-    '/team/edit/:id': EditTeam
+    '/team/edit/:id': TeamEdit
 }
